Move v2f model to separate files
Without semantic-aware merging, basic diff tools will confuse similar
functions or classes and think that changes to one are actually changes
to another.  This creates problems when merging with upstream code,
especially with the v2f turbulence model.  Since it is so similar to the
SST model, basic git merge or git diff tools will mess up the merge
completely unless manual editing is done.

By moving these classes and functions to their own files, we prevent
these diff-related mistakes entirely.

This should make merges much simpler.
