+++ SU2_CFD/include/solver_structure_v2f.hpp
@@ -1,0 +1,256 @@
+/*!
+ * \file solver_structure.hpp
+ * \brief Headers of the main subroutines for solving partial differential equations.
+ *        The subroutines and functions are in the <i>solver_structure.cpp</i>,
+ *        <i>solution_direct.cpp</i>, <i>solution_adjoint.cpp</i>, and
+ *        <i>solution_linearized.cpp</i> files.
+ * \author F. Palacios, T. Economon
+ * \version 6.0.1 "Falcon"
+ *
+ * The current SU2 release has been coordinated by the
+ * SU2 International Developers Society <www.su2devsociety.org>
+ * with selected contributions from the open-source community.
+ *
+ * The main research teams contributing to the current release are:
+ *  - Prof. Juan J. Alonso's group at Stanford University.
+ *  - Prof. Piero Colonna's group at Delft University of Technology.
+ *  - Prof. Nicolas R. Gauger's group at Kaiserslautern University of Technology.
+ *  - Prof. Alberto Guardone's group at Polytechnic University of Milan.
+ *  - Prof. Rafael Palacios' group at Imperial College London.
+ *  - Prof. Vincent Terrapon's group at the University of Liege.
+ *  - Prof. Edwin van der Weide's group at the University of Twente.
+ *  - Lab. of New Concepts in Aeronautics at Tech. Institute of Aeronautics.
+ *
+ * Copyright 2012-2018, Francisco D. Palacios, Thomas D. Economon,
+ *                      Tim Albring, and the SU2 contributors.
+ *
+ * SU2 is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * SU2 is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with SU2. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#pragma once
+
+#include "solver_structure.hpp"
+
+using namespace std;
+
+/*!
+ * \class CTurbKESolver
+ * \brief Main class for defining the turbulence model solver.
+ * \ingroup Turbulence_Model
+ * \author S. Haering
+ */
+
+class CTurbKESolver: public CTurbSolver {
+
+private:
+  su2double *constants,  /*!< \brief Constants for the model. */
+    kine_Inf,              /*!< \brief Free-stream turbulent kinetic energy. */
+    epsi_Inf,              /*!< \brief Free-stream specific dissipation. */
+    zeta_Inf,              /*!< \brief Free-stream v2/tke ratio. */
+    f_Inf;                 /*!< \brief Free-stream redistribution. */
+
+  /*!
+   * \brief Finish the averaging calculation.
+   *
+   * This is a templated step in the averaging calculation.  The averaging
+   * routine loops over all the nodes and calls this routine for each node.
+   *
+   * This step roughly corresponds to:
+   *   // Retrieve U_current
+   *   // Retrieve U_average
+   *   dU = (U_current - U_average)*weight;
+   *   // Store dU
+   *
+   * Note that the base class already updates the average of the solution.
+   * This method should only be implemented when other variables are to be
+   * averaged.
+   *
+   * \param weight - The amount to weight the update on the average
+   * \param iPoint - The point at which the average will be calculated
+   * \param buffer - An allocated array of size nVar for working calculations
+   */
+  void UpdateAverage(su2double weight, unsigned short iPoint, su2double* buffer);
+
+public:
+  /*!
+   * \brief Constructor of the class.
+   */
+  CTurbKESolver(void);
+
+  /*!
+   * \overload
+   * \param[in] geometry - Geometrical definition of the problem.
+   * \param[in] config - Definition of the particular problem.
+   * \param[in] iMesh - Index of the mesh in multigrid computations.
+   */
+  CTurbKESolver(CGeometry *geometry, CConfig *config, unsigned short iMesh);
+
+  /*!
+   * \brief Destructor of the class.
+   */
+  ~CTurbKESolver(void);
+
+  /*!
+   * \brief Restart residual and compute gradients.
+   * \param[in] geometry - Geometrical definition of the problem.
+   * \param[in] solver_container - Container vector with all the solutions.
+   * \param[in] config - Definition of the particular problem.
+   * \param[in] iMesh - Index of the mesh in multigrid computations.
+   * \param[in] iRKStep - Current step of the Runge-Kutta iteration.
+   * \param[in] RunTime_EqSystem - System of equations which is going to be solved.
+   * \param[in] Output - boolean to determine whether to print output.
+   */
+  void Preprocessing(CGeometry *geometry, CSolver **solver_container,
+                     CConfig *config,
+                     unsigned short iMesh, unsigned short iRKStep,
+                     unsigned short RunTime_EqSystem, bool Output);
+
+  /*!
+   * \brief Computes the eddy viscosity.
+   * \param[in] geometry - Geometrical definition of the problem.
+   * \param[in] solver_container - Container vector with all the solutions.
+   * \param[in] config - Definition of the particular problem.
+   * \param[in] iMesh - Index of the mesh in multigrid computations.
+   */
+  void Postprocessing(CGeometry *geometry, CSolver **solver_container,
+                      CConfig *config, unsigned short iMesh);
+  /*!
+   * \brief Calculates and stores the large-eddy timescale and lengthscale
+   * \param[in] solver_container - Container vector with all the solutions
+   * \param[in] config - Definition of the particular problem.
+   */
+  void CalculateTurbScales(CSolver **solver_container, CConfig *config);
+
+  /*!
+   * \brief Source term computation.
+   * \param[in] geometry - Geometrical definition of the problem.
+   * \param[in] solver_container - Container vector with all the solutions.
+   * \param[in] numerics - Description of the numerical method.
+   * \param[in] second_numerics - Description of the second numerical method.
+   * \param[in] config - Definition of the particular problem.
+   * \param[in] iMesh - Index of the mesh in multigrid computations.
+   */
+  void Source_Residual(CGeometry *geometry, CSolver **solver_container,
+                       CNumerics *numerics, CNumerics *second_numerics,
+                       CConfig *config, unsigned short iMesh,
+                       unsigned short iRKStep);
+
+  /*!
+   * \brief Source term computation.
+   * \param[in] geometry - Geometrical definition of the problem.
+   * \param[in] solver_container - Container vector with all the solutions.
+   * \param[in] numerics - Description of the numerical method.
+   * \param[in] config - Definition of the particular problem.
+   * \param[in] iMesh - Index of the mesh in multigrid computations.
+   */
+  void Source_Template(CGeometry *geometry, CSolver **solver_container,
+                       CNumerics *numerics, CConfig *config,
+                       unsigned short iMesh);
+
+  /*!
+   * \brief Impose the Navier-Stokes wall boundary condition.
+   * \param[in] geometry - Geometrical definition of the problem.
+   * \param[in] solver_container - Container vector with all the solutions.
+   * \param[in] conv_numerics - Description of the numerical method.
+   * \param[in] visc_numerics - Description of the numerical method.
+   * \param[in] config - Definition of the particular problem.
+   * \param[in] val_marker - Surface marker where the boundary condition is applied.
+   * \param[in] iRKStep - Current step of the Runge-Kutta iteration.
+   */
+  void BC_HeatFlux_Wall(CGeometry *geometry, CSolver **solver_container,
+                        CNumerics *conv_numerics, CNumerics *visc_numerics,
+                        CConfig *config, unsigned short val_marker,
+                        unsigned short iRKStep);
+
+  /*!
+   * \brief Impose the Navier-Stokes wall boundary condition.
+   * \param[in] geometry - Geometrical definition of the problem.
+   * \param[in] solver_container - Container vector with all the solutions.
+   * \param[in] conv_numerics - Description of the numerical method.
+   * \param[in] visc_numerics - Description of the numerical method.
+   * \param[in] config - Definition of the particular problem.
+   * \param[in] val_marker - Surface marker where the boundary condition is applied.
+   * \param[in] iRKStep - Current step of the Runge-Kutta iteration.
+   */
+  void BC_Isothermal_Wall(CGeometry *geometry, CSolver **solver_container,
+                          CNumerics *conv_numerics, CNumerics *visc_numerics,
+                          CConfig *config, unsigned short val_marker,
+                          unsigned short iRKStep);
+
+  /*!
+   * \brief Impose the Far Field boundary condition.
+   * \param[in] geometry - Geometrical definition of the problem.
+   * \param[in] solver_container - Container vector with all the solutions.
+   * \param[in] conv_numerics - Description of the numerical method.
+   * \param[in] visc_numerics - Description of the numerical method.
+   * \param[in] config - Definition of the particular problem.
+   * \param[in] val_marker - Surface marker where the boundary condition is applied.
+   * \param[in] iRKStep - Current step of the Runge-Kutta iteration.
+   */
+  void BC_Far_Field(CGeometry *geometry, CSolver **solver_container,
+                    CNumerics *conv_numerics, CNumerics *visc_numerics,
+                    CConfig *config, unsigned short val_marker,
+                    unsigned short iRKStep);
+
+  /*!
+   * \brief Impose the inlet boundary condition.
+   * \param[in] geometry - Geometrical definition of the problem.
+   * \param[in] solver_container - Container vector with all the solutions.
+   * \param[in] conv_numerics - Description of the numerical method.
+   * \param[in] visc_numerics - Description of the numerical method.
+   * \param[in] config - Definition of the particular problem.
+   * \param[in] val_marker - Surface marker where the boundary condition is applied.
+   * \param[in] iRKStep - Current step of the Runge-Kutta iteration.
+   */
+  void BC_Inlet(CGeometry *geometry, CSolver **solver_container,
+                CNumerics *conv_numerics, CNumerics *visc_numerics,
+                CConfig *config, unsigned short val_marker,
+                unsigned short iRKStep);
+
+  /*!
+   * \brief Impose the outlet boundary condition.
+   * \param[in] geometry - Geometrical definition of the problem.
+   * \param[in] solver_container - Container vector with all the solutions.
+   * \param[in] conv_numerics - Description of the numerical method.
+   * \param[in] visc_numerics - Description of the numerical method.
+   * \param[in] config - Definition of the particular problem.
+   * \param[in] val_marker - Surface marker where the boundary condition is applied.
+   * \param[in] iRKStep - Current step of the Runge-Kutta iteration.
+   */
+  void BC_Outlet(CGeometry *geometry, CSolver **solver_container,
+                 CNumerics *conv_numerics, CNumerics *visc_numerics,
+                 CConfig *config, unsigned short val_marker,
+                 unsigned short iRKStep);
+
+  /*!
+   * \brief Get the constants for the KE model.
+   * \return A pointer to an array containing a set of constants
+   */
+  su2double* GetConstants();
+
+  /*!
+   * \brief Set the solution using the Freestream values.
+   * \param[in] config - Definition of the particular problem.
+   */
+  void SetFreeStream_Solution(CConfig *config);
+
+
+  /*!
+   * \brief Setup the inlet per the config file and stores the result
+   * \param[in] config - Definition of the particular problem.
+   */
+  void SetInlet(CConfig *config);
+};
+
+#include "solver_structure_v2f.inl"
