+++ SU2_CFD/include/variable_structure.hpp
@@ -4647,94 +4647,6 @@ public:
   void SetRANSWeight(su2double val_w_rans);
 };
 
-/*! swh
- * \class CTurbKEVariable
- * \brief Main class for defining the variables of the turbulence model.
- * \ingroup Turbulence_Model
- * \author S. Haering
- * \version 4.3.x "Cardinal"
- */
-class CTurbKEVariable : public CTurbVariable {
-
-protected:
-  su2double sigma_e, sigma_k, sigma_z, C_e1o, C_e2, C1, C_2p, C_T, C_L, C_eta;
-  su2double Tm,		/*!< \brief T_m k-eps. */
-    Lm,		        /*!< \brief L_m k-eps */
-    Tm_avg,       /*!< \brief Average turbulent timescale */
-    Lm_avg,       /*!< \brief Average turbulent lengthscale */
-    Re_T;
-
-public:
-  /*!
-   * \brief Constructor of the class.
-   */
-  CTurbKEVariable(void);
-
-  /*!
-   * \overload
-   * \param[in] val_rho_kine - Turbulent variable value (initialization value).
-   * \param[in] val_rho_omega - Turbulent variable value (initialization value).
-   * \param[in] val_muT - Turbulent variable value (initialization value).
-   * \param[in] val_nDim - Number of dimensions of the problem.
-   * \param[in] val_nvar - Number of variables of the problem.
-   * \param[in] constants -
-   * \param[in] config - Definition of the particular problem.
-   */
-  CTurbKEVariable(su2double val_rho_kine, su2double val_rho_epsi,
-                  su2double val_zeta, su2double val_f,
-                  su2double val_muT, su2double val_Tm, su2double val_Lm,
-                  unsigned short val_nDim, unsigned short val_nvar,
-                  su2double *constants, CConfig *config);
-
-  /*!
-   * \brief Destructor of the class.
-   */
-  ~CTurbKEVariable(void);
-
-  /**
-   * \brief Get the large-eddy timescale of the turbulence
-   * \return The large-eddy timescale of the turbulence.
-   */
-  su2double GetTurbTimescale(void);
-
-  /**
-   * \brief Get the large-eddy lengthscale of the turbulence
-   * \return The large-eddy lengthscale of the turbulence
-   */
-  su2double GetTurbLengthscale(void);
-
-  /**
-   * \brief Get the average large-eddy timescale of the turbulence
-   * \return The large-eddy timescale of the turbulence.
-   */
-  su2double GetAverageTurbTimescale(void);
-
-  /**
-   * \brief Get the average large-eddy lengthscale of the turbulence
-   * \return The large-eddy lengthscale of the turbulence
-   */
-  su2double GetAverageTurbLengthscale(void);
-
-  /*!
-   * \brief Get the component anisotropy ratio (max-to-min)
-   * \return The Reynolds stress component anisotropy ratio (max-to-min)
-   */
-  su2double GetAnisoRatio(void);
-
-  /**
-   * \brief Sets the large-eddy lengthscale and the large-eddy timescale
-   * \param[in] val_turb_T - Large eddy timescale of the turbulence
-   * \param[in] val_turb_L - Large eddy lengthscale of the turbulence
-   */
-  void SetTurbScales(su2double val_turb_T, su2double val_turb_L);
-
-  /**
-   * \brief Sets the average large-eddy lengthscale and the timescale
-   * \param[in] val_T_avg - Average large eddy timescale of the turbulence
-   * \param[in] val_L_avg - Average large eddy lengthscale of the turbulence
-   */
-  void SetAverageTurbScales(su2double val_T_avg, su2double val_L_avg);
-};
 
 
 /*!
