+++ SU2_CFD/include/variable_structure_v2f.inl
@@ -1,0 +1,65 @@
+/*!
+ * \file variable_structure.inl
+ * \brief In-Line subroutines of the <i>variable_structure.hpp</i> file.
+ * \author F. Palacios, T. Economon
+ * \version 6.0.1 "Falcon"
+ *
+ * The current SU2 release has been coordinated by the
+ * SU2 International Developers Society <www.su2devsociety.org>
+ * with selected contributions from the open-source community.
+ *
+ * The main research teams contributing to the current release are:
+ *  - Prof. Juan J. Alonso's group at Stanford University.
+ *  - Prof. Piero Colonna's group at Delft University of Technology.
+ *  - Prof. Nicolas R. Gauger's group at Kaiserslautern University of Technology.
+ *  - Prof. Alberto Guardone's group at Polytechnic University of Milan.
+ *  - Prof. Rafael Palacios' group at Imperial College London.
+ *  - Prof. Vincent Terrapon's group at the University of Liege.
+ *  - Prof. Edwin van der Weide's group at the University of Twente.
+ *  - Lab. of New Concepts in Aeronautics at Tech. Institute of Aeronautics.
+ *
+ * Copyright 2012-2018, Francisco D. Palacios, Thomas D. Economon,
+ *                      Tim Albring, and the SU2 contributors.
+ *
+ * SU2 is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * SU2 is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with SU2. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#pragma once
+
+inline su2double CTurbKEVariable::GetTurbTimescale(){
+  return Tm;
+}
+
+inline su2double CTurbKEVariable::GetTurbLengthscale() {
+ return Lm;
+}
+
+inline su2double CTurbKEVariable::GetAverageTurbTimescale() {
+  return Tm_avg;
+}
+
+inline su2double CTurbKEVariable::GetAverageTurbLengthscale() {
+ return Lm_avg;
+}
+
+inline void CTurbKEVariable::SetTurbScales(su2double val_turb_T, su2double val_turb_L) {
+  Tm = val_turb_T;
+  Lm = val_turb_L;
+}
+
+inline void CTurbKEVariable::SetAverageTurbScales(su2double val_T_avg, su2double val_L_avg) {
+  Tm_avg = val_T_avg;
+  Lm_avg = val_L_avg;
+}
+
