+++ SU2_CFD/src/driver_structure.cpp
@@ -37,6 +37,9 @@
 
 #include "../include/driver_structure.hpp"
 #include "../include/definition_structure.hpp"
+#include "../include/variable_structure_v2f.hpp"
+#include "../include/numerics_structure_v2f.hpp"
+#include "../include/solver_structure_v2f.hpp"
 
 CDriver::CDriver(char* confFile,
                  unsigned short val_nZone,
