+++ SU2_CFD/src/variable_direct_turbulent.cpp
@@ -241,50 +241,3 @@ void CTurbSSTVariable::SetBlendingFunc(su2double val_viscosity, su2double val_di
   AD::EndPreacc();
   
 }
-
-// swh
-CTurbKEVariable::CTurbKEVariable(void) : CTurbVariable() { }
-
-CTurbKEVariable::CTurbKEVariable(su2double val_kine, su2double val_epsi,
-                                 su2double val_zeta, su2double val_f,
-                                 su2double val_muT, su2double val_Tm,
-                                 su2double val_Lm, unsigned short val_nDim,
-                                 unsigned short val_nvar,
-                                 su2double *constants, CConfig *config)
-  :
-  CTurbVariable(val_nDim, val_nvar, config) {
-
-  bool dual_time = ((config->GetUnsteady_Simulation() == DT_STEPPING_1ST) ||
-                    (config->GetUnsteady_Simulation() == DT_STEPPING_2ND));
-
-  /*--- Initialization of variables ---*/
-  Solution[0] = val_kine; Solution_Old[0] = val_kine;
-  Solution[1] = val_epsi;	Solution_Old[1] = val_epsi;
-  Solution[2] = val_zeta; Solution_Old[2] = val_zeta;
-  Solution[3] = val_f;	Solution_Old[3] = val_f;
-  Tm  = val_Tm;
-  Lm  = val_Lm;
-  Tm_avg = val_Tm;
-  Lm_avg = val_Lm;
-
-  /*--- Initialization of eddy viscosity ---*/  
-  muT = val_muT;
-
-  /*--- Allocate and initialize solution for the dual time strategy ---*/
-  if (dual_time) {
-    Solution_time_n[0]  = val_kine; Solution_time_n[1]  = val_epsi;
-    Solution_time_n1[0] = val_kine; Solution_time_n1[1] = val_epsi;
-    Solution_time_n[2]  = val_zeta; Solution_time_n[3]  = val_f;
-    Solution_time_n1[2] = val_zeta; Solution_time_n1[3] = val_f;
-  }
-
-}
-
-CTurbKEVariable::~CTurbKEVariable(void) {
-}
-
-su2double CTurbKEVariable::GetAnisoRatio(void) {
-  // XXX: This floor is arbitrary.
-  const su2double TKE_MIN = EPS;
-  return TWO3*Solution[0]/max(TKE_MIN, Solution[2]);
-}
